[FIX] docker-odoo-image: Fix #135 the abscence of vauxoo 9.0 branch
Instead of using a cloning for a specific branch version, we add an empty folder
to transform it in a .git repo and then we add the necessary remotes.

With this change, is allowed to fetch the vauxoo 8.0 or vauxoo 9.0 branch without the
fetching problems at using git clone -b 8.0 --single-branch, because that command only
fetch the history of the specific branch and not allows to integrate the other branches.