[FIX] Install dev packages for postgres to avoid the exception.
We need to install postgresql-server-dev so the pg_config commands shows the proper version, at the moment it shows 10.0 even when only the 9.6 is installed. This issue was fixed in pycopg 2.7, but not in the previous versions: https://github.com/psycopg/psycopg2/issues/594

This patch only apply for the images that uses older pycopg versions, that is 8.0 and 9.0 because 10 and 11 use 2.7.