[IMP] odoo-shippable: Apply VX's custom config for all PostgreSQL
Currently, custom settings for PostgreSQL, available on the image
`docker-ubuntu-base`, were not applied in the big image. One of the
consequences of that is the parameter `autovacuum` was not being
disabled.

This commit causes the Vauxoo's custom tune-ups to be applied in the big
image, for all installed versions.

In addition, this removes the following customizations:
- Custom configuration of max connections, port and locks:
  these were not taken into account, because all config files are removed
  after those settings are applied. To be more specific, this happens
  because the function `service_postgres_without_sudo` is called after the
  aforementioned settings are applied, which run `pg_dropcluster` for all
  versions, removing all previously-modified config files. In addition,
  they are included in:
  https://github.com/Vauxoo/docker-ubuntu-base/pull/41
- Enabling PG LOGS AND NON DURABILITY:
  These and other ones are already enabled thanks to this commit, so there
  is no need for setting up them twice.

Closes #305