[FIX] build.py: Avoid false positives when build fails
Before this commit, the return value of the command `docker build` was
not being taken into account, thus producing false positives when the
build failed.

After this commit, the return value is taken into account and it's
returned, causing the script to return a non-zero value when the build
command returns a non-zero value.