[FIX] odoo-shippable: Point PostgreSQL logs correctly on aliases
After Vauxoo's config settings for PostgreSQL were applied on #308,
PostgreSQL's logs directory changed, from within the cluster directory
to `/var/log`. Because of that, aliases referring to the old location
needs to be updated.

This sets correctly log paths in the aliases psql_log_*.

In addition, this avoids inserting an extra newline ad beginning of file
when logs are cleared.

Closes https://github.com/Vauxoo/docker-odoo-image/issues/322